[CL 3228288 by Peter Knepley in UT-Clean branch]

+++ Engine/Source/Developer/HTML5/HTML5TargetPlatform/Private/HTML5TargetDevice.cpp
@@ -1,6 +1,9 @@
 // Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
 
-#include "HTML5TargetPlatformPrivatePCH.h"
+#include "HTML5TargetDevice.h"
+#include "Interfaces/ITargetPlatform.h"
+
+
 
 /* ITargetDevice interface
  *****************************************************************************/
@@ -105,8 +108,7 @@ bool FHTML5TargetDevice::GetUserCredentials( FString& OutUserName, FString& OutU
 	return false;
 }
 
-bool FHTML5TargetDevice::TerminateProcess( const int32 ProcessId )
+bool FHTML5TargetDevice::TerminateProcess( const int64 ProcessId )
 {
 	return false;
 }
-
