+++ Engine/Source/Developer/iOS/IOSTargetPlatform/Private/IOSTargetDevice.cpp
@@ -1,6 +1,7 @@
 // Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
 
-#include "IOSTargetPlatformPrivatePCH.h"
+#include "IOSTargetDevice.h"
+#include "IOSMessageProtocol.h"
 
 bool FIOSTargetDevice::Connect()
 {
@@ -154,7 +155,7 @@ bool FIOSTargetDevice::SupportsSdkVersion(const FString& VersionString) const
 	return true;
 }
 
-bool FIOSTargetDevice::TerminateProcess(const int32 ProcessId)
+bool FIOSTargetDevice::TerminateProcess(const int64 ProcessId)
 {
 	return false;
 }
