+++ Engine/Source/Developer/iOS/IOSTargetPlatform/Private/IOSTargetDevice.h
@@ -6,6 +6,14 @@
 
 #pragma once
 
+#include "CoreMinimal.h"
+#include "Interfaces/ITargetDevice.h"
+#include "Interfaces/ITargetPlatform.h"
+#include "HAL/PlatformProcess.h"
+#include "IMessageContext.h"
+#include "Helpers/MessageEndpoint.h"
+#include "Helpers/MessageEndpointBuilder.h"
+
 
 /**
  * Type definition for shared pointers to instances of FIOSTargetDevice.
@@ -57,7 +65,7 @@ public:
 	virtual bool Run(const FString& ExecutablePath, const FString& Params, uint32* OutProcessId) override;
 	virtual bool SupportsFeature(ETargetDeviceFeatures Feature) const;
 	virtual bool SupportsSdkVersion(const FString& VersionString) const override;
-	virtual bool TerminateProcess(const int32 ProcessId) override;
+	virtual bool TerminateProcess(const int64 ProcessId) override;
 	virtual void SetUserCredentials(const FString& UserName, const FString& UserPassword) override;
 	virtual bool GetUserCredentials(FString& OutUserName, FString& OutUserPassword) override;
 	// End of ITargetDevice interface
